fix import contacts

+++ services/sim/src/usim_dialling_numbers_service.cpp
@@ -107,7 +107,8 @@ void UsimDiallingNumbersService::ProcessDiallingNumberLoadDone(const AppExecFwk:
         LoadDiallingNumber2Files(currentIndex_);
         return;
     }
-    TELEPHONY_LOGI("usimservice load adn done, fileId=%{public}#llX", resultObject->fileID);
+    TELEPHONY_LOGI("usimservice load adn done, fileId=%{public}#llX",
+        static_cast<unsigned long long>(resultObject->fileID));
     adns_[resultObject->fileID] = std::vector<std::shared_ptr<DiallingNumbersInfo>>();
     if (resultObject->exception != nullptr) {
         auto exception = std::static_pointer_cast<RadioResponseInfo>(resultObject->exception);
